KubeVc/requirements.txt
Flask
boto3==1.28.52
PyYAML
kubernetes
diagrams
graphviz

